requirement.txt
Flask==2.3.2
Flask-WTF==1.2.1
WTForms==3.0.1
python-dotenv==1.0.0
email-validator==2.1.0
markupsafe==2.1.3
